% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderSafePivot}
\alias{renderSafePivot}
\title{Shiny render function for safePivot}
\usage{
renderSafePivot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a safePivot widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is already quoted.}
}
\value{
A Shiny render function.
}
\description{
Shiny render function for safePivot
}
