% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{safe_pivot_write_rds}
\alias{safe_pivot_write_rds}
\title{Write pivot result to RDS}
\usage{
safe_pivot_write_rds(pivot_result, file)
}
\arguments{
\item{pivot_result}{A result object returned by \code{safe_pivot_compute()}.}

\item{file}{Output \code{.rds} file path.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Write pivot result to RDS
}
