% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{safePivotOutput}
\alias{safePivotOutput}
\title{Shiny output binding for safePivot}
\usage{
safePivotOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Output variable name.}

\item{width}{Width of the widget. e.g. "100\%", "400px" or "80vh"}

\item{height}{Height of the widget.}
}
\value{
A Shiny widget output object.
}
\description{
Shiny output binding for safePivot
}
